requirements.txt
streamlit
openai
chromadb
pdfplumber
python-docx
tqdm
pandas
python-dotenv

